Finalized files work
-Added global.h: has what struct had, plus everything that was global in main.c, also has all the #defines, and the #includes of standart c libraries, so the other files don't need to redefine or include anything
-Placed every function in main (except for main itself) in its respective file
