+++ enemies.h
@@ -7,7 +7,7 @@
 #include <time.h>
 // CSFML header
 #include <SFML/Graphics.h>
-#include "structs.h"
+#include "global.h"
 
 // This creates one enemy and returns it.
 TYPE_ENEMIES Enemy_Create(int color, int posX, int posY);
@@ -29,5 +29,7 @@ int Enemies_HowManyFires(TYPE_ENEMIES enemies[], int numberEnemies);
 int Enemies_canShoot(TYPE_ENEMIES enemies[], int numberEnemies);
 // This destroys all of the enemies' fires
 void Enemies_destroyFires(TYPE_ENEMIES enemies[], int numberEnemies);
+// This returns the number of enemy in the array, if the sprite given has the same position than that enemy. Else, it returns -1.
+int isAtSamePoint(TYPE_ENEMIES* enemies, int sizeArray, sfSprite* sprite);
 
 #endif // ENEMIES_H_INCLUDED
