+++ global.c
@@ -1,0 +1,7 @@
+#include "global.h"
+
+float energy = ENERGYMAX;
+int numberlifes = 3;
+int nEnemies = 0;
+int liveEnemies = 0;
+int score = 0;
